+++ prevedere/prevedere.py
@@ -1,6 +1,4 @@
 import requests
-import pandas as pd
-from pandas.io.json import json_normalize
 
 
 class Api:
@@ -12,16 +10,20 @@ class Api:
     def fetch(self, path: str, payload: dict = {}) -> dict:
         url = f'https://api.prevedere.com{path}'
         payload['ApiKey'] = self.api_key
-        r = requests.get(url, params=payload)
-        return r.json()
-    
+
+        try:
+            r = requests.get(url, params=payload)
+            r.raise_for_status()
+            return r.json()
+        except requests.exceptions.RequestException as err:
+            print(err)
+
     def indicator_series(self,
                          provider: str,
                          provider_id: str,
                          freq: str = 'Monthly',
                          calculation: str = None,
-                         offset: int = 0,
-                         verbose: bool = False) -> pd.DataFrame:
+                         offset: int = 0) -> dict:
         """
         Create a pandas dataframe with data from the API.
 
@@ -36,55 +38,30 @@ class Api:
         :type freq: str
 
         :param calculation: Calculation to transform the indicator
-        ("None","PeriodOverPeriod","YearOverYear","ThreePeriodMoving","FivePeriodMoving","ThreePeriodYearOverYear")
+        ("None","PeriodOverPeriod","YearOverYear","ThreePeriodMoving",
+        "FivePeriodMoving","ThreePeriodYearOverYear")
         :type calculation: string
 
         :param offset: Number of periods to offset
         :type offset: int
-
-        :param verbose:
-        :type verbose: bool
         """
+
         path = f'/indicator/series/{provider}/{provider_id}'
         payload = {'Frequency': freq, 'Offset': offset, "Calculation": calculation}
-        df = pd.DataFrame(self.fetch(path, payload))
-        df.columns = df.columns.str.lower()
+        return self.fetch(path, payload)
 
-        if verbose:
-            print(self.indicator(provider, provider_id).loc[
-                      ['ProviderId',
-                       'Provider.Name',
-                       'Name',
-                       'Source',
-                       'Units',
-                       'Frequency',
-                       'StartTime',
-                       'EndTime']
-                  ])
-
-        if "date" in df.columns:
-            df['date'] = pd.to_datetime(df['date'])
-            df.set_index("date", inplace=True)
-
-        return df
-
-    def indicator(self, provider: str, provider_id: str) -> pd.DataFrame:
+    def indicator(self, provider: str, provider_id: str) -> dict:
         path = f'/indicator/{provider}/{provider_id}'
-        return json_normalize(self.fetch(path)).T
-    
-    def providers(self) -> dict:
-        path = '/providers'
         return self.fetch(path)
 
-    def analysis_jobs(self) -> dict:
-        path = '/analysisjobs'
+    def providers(self) -> dict:
+        path = '/provider'
         return self.fetch(path)
-    
-    def search(self, query: str) -> pd.DataFrame:
+
+    def search(self, query: str) -> dict:
         path = '/search'
         payload = {'Query': query}
-        results = self.fetch(path, payload)
-        return json_normalize(data=results['Results'])
+        return self.fetch(path, payload)
 
     def model(self, model_id: str) -> dict:
         path = f'/rawmodel/{model_id}'
@@ -97,4 +74,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-